+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix background tasks redis
+
 ### 0.11.3 - 2019-10-24
 
  - Fix background tasks